refactor(test): move findElementByType to src/__tests__/utils; merge OverallContent seed tests (PR #527 review)
