+++ src/widgets/overall/__tests__/OverallContent.test.tsx
@@ -7,6 +7,10 @@ import type { MockedFunction } from 'vitest';
 
 // vi.mock은 vitest가 import 위로 hoist하지만, ESLint(import/first)와
 // 가독성을 위해 소스 코드에서도 모든 import보다 위에 둔다.
+//
+// 대부분의 describe는 useOverallAnalysis를 mock해 임의 state shape를 주입하지만,
+// 'OverallContent SSR seed' describe는 mockImplementation으로 실제 훅을 복원해
+// initialAnalysis prop → done 서사로 흐르는 전체 경로를 검증한다(아래 참고).
 vi.mock('@/widgets/overall/hooks/useOverallAnalysis', () => ({
     useOverallAnalysis: vi.fn(),
 }));
@@ -24,16 +28,41 @@ vi.mock('@/shared/ui/MarkdownText', () => ({
         <div>{children}</div>
     ),
 }));
+// Server Action 묶음 — SSR seed describe가 실제 useOverallAnalysis를 사용할 때
+// 훅이 import하는 server-only 체인을 끊기 위해 mock한다. seed(done) 경로에서는
+// submitOverallAnalysisAction이 호출되지 않음을 검증한다.
+vi.mock('@/entities/analysis/actions', () => ({
+    submitOverallAnalysisAction: vi.fn(),
+    pollOverallAnalysisAction: vi.fn(),
+    pollAnalysisAction: vi.fn(),
+    pollFundamentalAnalysisAction: vi.fn(),
+    cancelAnalysisJobAction: vi.fn().mockResolvedValue(undefined),
+    cancelFundamentalAnalysisJobAction: vi.fn().mockResolvedValue(undefined),
+    cancelOverallAnalysisJobAction: vi.fn().mockResolvedValue(undefined),
+}));
+vi.mock('@/entities/news-article/actions', () => ({
+    pollNewsAnalysisAction: vi.fn(),
+    cancelNewsAnalysisJobAction: vi.fn().mockResolvedValue(undefined),
+}));
+vi.mock('@/entities/options-chain/actions', () => ({
+    pollOptionsAnalysisAction: vi.fn(),
+    cancelOptionsAnalysisJobAction: vi.fn().mockResolvedValue(undefined),
+}));
 
 import { render, screen, fireEvent } from '@testing-library/react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import type { ReactNode } from 'react';
 import type { OverallAnalysisResponse } from '@y0ngha/siglens-core';
 
 import { OverallContent } from '@/widgets/overall/OverallContent';
 import { useOverallAnalysis } from '@/widgets/overall/hooks/useOverallAnalysis';
+import { submitOverallAnalysisAction } from '@/entities/analysis/actions';
 
 const mockUseOverallAnalysis = useOverallAnalysis as MockedFunction<
     typeof useOverallAnalysis
+>;
+const mockSubmit = submitOverallAnalysisAction as MockedFunction<
+    typeof submitOverallAnalysisAction
 >;
 
 function makeDoneResult(
@@ -329,5 +358,82 @@ describe('OverallContent done branch', () => {
         );
         fireEvent.click(screen.getByRole('button', { name: /재분석/ }));
         expect(trigger).toHaveBeenCalledTimes(1);
+    });
+});
+
+/**
+ * OverallContent SSR seed 테스트. 위 describe들과 달리 useOverallAnalysis를
+ * mock하지 않고 (mockImplementation으로 실제 훅을 복원해) initialAnalysis prop이
+ * seed(done 상태)로 흐르는 전체 경로를 검증한다. Server Action은 모듈 상단에서
+ * mock해 네트워크/server-only 의존을 차단한다.
+ */
+const SEED_RESULT: OverallAnalysisResponse = {
+    headlineKo: 'AAPL 시드 헤드라인',
+    technicalBulletsKo: ['기술적 신호'],
+    fundamentalBulletsKo: ['펀더멘털 신호'],
+    newsBulletsKo: ['뉴스 신호'],
+    optionsBulletsKo: ['옵션 신호'],
+    integratedConclusionKo: '통합 결론',
+    scenarios: [],
+    riskFactorsKo: [],
+};
+
+function makeWrapper() {
+    const client = new QueryClient({
+        defaultOptions: { queries: { retry: false } },
+    });
+    return function Wrapper({ children }: { children: ReactNode }) {
+        return (
+            <QueryClientProvider client={client}>
+                {children}
+            </QueryClientProvider>
+        );
+    };
+}
+
+describe('OverallContent SSR seed', () => {
+    beforeEach(async () => {
+        mockSubmit.mockReset();
+        // 실제 훅을 복원해 initialAnalysis → query.initialData → done 경로를 그대로 탄다.
+        const actual = await vi.importActual<
+            typeof import('@/widgets/overall/hooks/useOverallAnalysis')
+        >('@/widgets/overall/hooks/useOverallAnalysis');
+        mockUseOverallAnalysis.mockImplementation(actual.useOverallAnalysis);
+    });
+
+    afterEach(() => {
+        // 다른 describe는 mockReturnValue로 state를 강제 주입하므로 실제 구현 복원이
+        // 누수되지 않도록 reset한다.
+        mockUseOverallAnalysis.mockReset();
+    });
+
+    it('initialAnalysis가 주어지면 done 서사(headline)를 즉시 렌더하고 생성을 트리거하지 않는다', () => {
+        render(
+            <OverallContent
+                symbol="AAPL"
+                companyName="Apple Inc."
+                timeframe="1Day"
+                initialAnalysis={SEED_RESULT}
+            />,
+            { wrapper: makeWrapper() }
+        );
+
+        expect(screen.getByText('AAPL 시드 헤드라인')).toBeInTheDocument();
+        expect(mockSubmit).not.toHaveBeenCalled();
+    });
+
+    it('initialAnalysis가 없으면 idle CTA(분석 받기)를 렌더한다', () => {
+        render(
+            <OverallContent
+                symbol="AAPL"
+                companyName="Apple Inc."
+                timeframe="1Day"
+            />,
+            { wrapper: makeWrapper() }
+        );
+
+        expect(
+            screen.getByRole('button', { name: /AI 종합 분석 받기/ })
+        ).toBeInTheDocument();
     });
 });
